Tests for scikit time objects
